Session keyword PLUGIN added. Given plugins now get downloaded to session_plugin (specified in Configuration) and loaded.

+++ src/net/sf/genomeview/plugin/PluginLoader.java
@@ -6,8 +6,8 @@ package net.sf.genomeview.plugin;
 import java.io.File;
 import java.io.FilenameFilter;
 import java.io.IOException;
+import java.net.URISyntaxException;
 import java.net.URL;
-import java.net.URLConnection;
 import java.util.Collection;
 import java.util.HashSet;
 import java.util.Iterator;
@@ -18,6 +18,7 @@ import javax.swing.JOptionPane;
 
 import net.sf.genomeview.core.Configuration;
 import net.sf.genomeview.data.Model;
+import net.sf.jannot.source.Locator;
 
 import org.apache.commons.io.FileUtils;
 import org.java.plugin.ObjectFactory;
@@ -271,6 +272,22 @@ public class PluginLoader {
 	/**
 	 * Gets a plugin location, copies the plugin to the given destination directory and registers/activates the new plugin with 
 	 * {@link #loadPlugin(File)}.
+	 * If the plugin already existed in the target folder, it will not be overwritten.
+	 * 
+	 * @param pluginLocation place download the jar/zip file from 
+	 * @param pluginDirectory place to store the plugin. This will be eiter the default plugin directory or the sessions plugin dir.
+	 */
+	public static void installPlugin(Locator pluginLocation, File pluginDirectory) throws IOException, URISyntaxException{
+		if (pluginLocation.isURL()){
+			installPlugin(pluginLocation.url(), pluginDirectory);
+		} else {
+			installPlugin(pluginLocation.file(), pluginDirectory);
+		}
+	}
+	/**
+	 * Gets a plugin location, copies the plugin to the given destination directory and registers/activates the new plugin with 
+	 * {@link #loadPlugin(File)}.
+	 * If the plugin already existed in the target folder, it will not be overwritten.
 	 * 
 	 * @param pluginLocation place download the jar/zip file from 
 	 * @param pluginDirectory place to store the plugin. This will be eiter the default plugin directory or the sessions plugin dir.
@@ -279,13 +296,16 @@ public class PluginLoader {
 		String[] path = pluginLocation.getPath().split("/");
 		String jarName = pluginLocation.getPath().split("/")[path.length-1];
 		File dest = new File(pluginDirectory, jarName);
-		FileUtils.copyURLToFile(pluginLocation, dest);
+		if (!dest.exists()){
+			FileUtils.copyURLToFile(pluginLocation, dest);			
+		}
 		
 		loadPlugin(dest);
 	}
 	/**
 	 * Gets a plugin location, copies the plugin to the given destination directory and registers/activates the new plugin with 
 	 * {@link #loadPlugin(File)}.
+	 * If the plugin already existed in the target folder, it will not be overwritten.
 	 * 
 	 * @param pluginLocation place to copy the jar/zip file from 
 	 * @param pluginDirectory place to store the plugin. This will be eiter the default plugin directory or the sessions plugin dir.
@@ -294,7 +314,9 @@ public class PluginLoader {
 		String[] path = pluginLocation.getPath().split("/");
 		String jarName = pluginLocation.getPath().split("/")[path.length-1];
 		File dest = new File(pluginDirectory, jarName);
-		FileUtils.copyFile(pluginLocation, dest);
+		if (!dest.exists()){
+			FileUtils.copyFile(pluginLocation, dest);			
+		}
 		
 		loadPlugin(dest);
 	}
